+++ src/discovery/index.ts
@@ -1,5 +1,6 @@
 // Copyright 2026 Antifraud Services Inc. under the Apache License, Version 2.0.
 
 export * from './DataspaceResolver.js';
+export * from './DefaultFirstDataspaceDiscovery.js';
 export * from './HttpDataspaceResolver.js';
 export * from './types.js';